Changed search bar on Main Screen
Search bar is now drop-down menu of the beaches found in the rtdb. clicking on a beach will automatically take you to the weather screen. BeachID is now found in Main Screen and is just an app variable that initializes null.

clicking the beaches button brings down the list of beaches, clicking it again puts it away.

Everything is put into /srcfixed